Added the notes and made search work

+++ src/main/java/com/theironyard/entity/Search.java
@@ -4,17 +4,17 @@ package com.theironyard.entity;
  * Created by katherine_celeste on 10/4/16.
  */
 public class Search {
-    Integer typeId;
+    Long typeId;
     String name;
 
     public Search() {
     }
 
-    public Integer getTypeId() {
+    public Long getTypeId() {
         return typeId;
     }
 
-    public void setTypeId(Integer typeId) {
+    public void setTypeId(Long typeId) {
         this.typeId = typeId;
     }
